+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Updated fall damage values for v49
+- Fixed immediate fall damage when using the jetpack for the first time while moving fast enough
+
 ## 1.0.1
 
 - Updated readme, no code changes